fixed some bugs
the jump instruction: there was an error where the PC_JUMP wasn't written properly as we added 2 zero bits then we shifted the immediate address so it turned like that we shifted two times, the other thing is that we need to store the 26 bits of the immediate address in a 28-bits storage in order to perform proper shifting by 2. Also, the regwrite wasn't connected to WE3 in regfile and reset in data memory!